#250 [REFACTOR] visitInfoResponse 값을 순차적으로 변경하기 위해, record 였던 VisitInfoResponse 를 class 로 변경 및 다른 response 들도 statistics 로 디렉토리 변경
