+++ src/main/java/com/developlife/reviewtwits/controller/StatController.java
@@ -3,9 +3,9 @@ package com.developlife.reviewtwits.controller;
 import com.developlife.reviewtwits.entity.User;
 import com.developlife.reviewtwits.message.annotation.project.ChartPeriod;
 import com.developlife.reviewtwits.message.request.StatMessageRequest;
-import com.developlife.reviewtwits.message.response.project.DailyVisitInfoResponse;
+import com.developlife.reviewtwits.message.response.statistics.DailyVisitInfoResponse;
 import com.developlife.reviewtwits.message.response.project.RecentVisitInfoResponse;
-import com.developlife.reviewtwits.message.response.project.VisitTotalGraphResponse;
+import com.developlife.reviewtwits.message.response.statistics.VisitTotalGraphResponse;
 import com.developlife.reviewtwits.message.response.statistics.SaveStatResponse;
 import com.developlife.reviewtwits.service.StatService;
 import lombok.RequiredArgsConstructor;
