+++ src/main/java/com/developlife/reviewtwits/service/StatService.java
@@ -9,10 +9,10 @@ import com.developlife.reviewtwits.exception.project.ProjectNotFoundException;
 import com.developlife.reviewtwits.exception.user.AccessDeniedException;
 import com.developlife.reviewtwits.mapper.StatMapper;
 import com.developlife.reviewtwits.message.request.StatMessageRequest;
-import com.developlife.reviewtwits.message.response.project.DailyVisitInfoResponse;
+import com.developlife.reviewtwits.message.response.statistics.DailyVisitInfoResponse;
 import com.developlife.reviewtwits.message.response.project.RecentVisitInfoResponse;
-import com.developlife.reviewtwits.message.response.project.VisitInfoResponse;
-import com.developlife.reviewtwits.message.response.project.VisitTotalGraphResponse;
+import com.developlife.reviewtwits.message.response.statistics.VisitInfoResponse;
+import com.developlife.reviewtwits.message.response.statistics.VisitTotalGraphResponse;
 import com.developlife.reviewtwits.message.response.statistics.SaveStatResponse;
 import com.developlife.reviewtwits.repository.ProductRepository;
 import com.developlife.reviewtwits.repository.project.ProjectRepository;
